cortex_engine/requirements.txt
fastapi
uvicorn
httpx
numpy
scikit-learn
pandas
openai
python-dotenv
# Snowflake dependencies for hybrid system
snowflake-connector-python
snowflake-sqlalchemy
sqlalchemy
# Testing dependencies
pytest
